add apis

+++ ToDoApp/alembic/versions/491cc784fff8_create_phone_number_for_user_column.py
@@ -20,9 +20,17 @@ depends_on: Union[str, Sequence[str], None] = None
 
 def upgrade() -> None:
     """Upgrade schema."""
-    op.add_column('users', sa.Column('phoneNumber', sa.String(), nullable=True))
+    conn = op.get_bind()
+    inspector = sa.inspect(conn)
+    columns = {col["name"] for col in inspector.get_columns("users")}
+    if "phoneNumber" not in columns:
+        op.add_column('users', sa.Column('phoneNumber', sa.String(), nullable=True))
 
 
 def downgrade() -> None:
     """Downgrade schema."""
-    op.drop_column('users', 'phoneNumber')
+    conn = op.get_bind()
+    inspector = sa.inspect(conn)
+    columns = {col["name"] for col in inspector.get_columns("users")}
+    if "phoneNumber" in columns:
+        op.drop_column('users', 'phoneNumber')
